Introduce new SourceLookUpProvider
The new SourceLookUpProvider is built from a  Seq[ClassPathEntry]
that must be provided by the client (the build server).

A ClassPathEntry is composed of:
- the path of a java class folder or class jar
- its corresponding source entries: a source jar, a source directory,
a bunch of source files

For each ClassPathEntry, the SourceLookupProvider creates a mapping
from class files to source files by reading the source name of each
class file. When a source name matches several source files, we cannot
be certain which one is correct. We try to recover from that
situation by checking if the class package matches the relative path of
the souce file. This is only convention.

If we do not find a source file we say that the class file is an orphan,
which is only problematic if the class file contains some line numbers.

The ClassPathEntryLookUpStats is the "test" class that prints
the loading time and the number of orphan classes of some big Scala
libraries (artifact).