Prevent object reference from being collected
Objects created on the remote JVM can be garbage-collected at any time.
So during expression evaluation, we need to protect those by calling
`disableCollection` once. We also need to call `enableCollection`
after the expression evaluation succeeded or failed.

For reference, see https://stackoverflow.com/questions/25793688/life-span-of-jdi-mirrors-of-objects-living-in-a-remote-jvm

To do so I introduce a `Safe[T]` class which encapsulates the result of
a computation on remote objects. All remote are protected until one
calls `safe.get`, which enables the collection once again.