Fix evalauting in tail recursive annotated functions
Previously, when trying to evaluate tail recursive function inside itself the compilation of the expression might fail due to the function no longer being tail recursive. Now, we remove the annotation when generating the evaluation and also do not break when evaluating, which would cause the function to stop again.

I also added pprint for nicer debugging messages, but it's not 100% neccessary.