Matches types of args and result to handle overloading
This is to avoid stepping into a mixin-forwarder or a bridge because
we matchs one of its overload.

The matching on types is a best effort because we don't access the
external type aliases:
- We try to encode all external types based on
their path.
- If this encoded type is known by the class loader, that
means we successfully encoded it.
- If it is not known by the classloader,
there must be a type parameter or a type alias in a path, but we don't
try to resolve it.

Type aliases of the Scala lib are so pervasive that we encoded them statically,
to reduce the number of false positives.