Make `matchType` more precise by directly working on ErasedTypeRef's.
And by explicitly following owners one at a time, instead of
converting the whole signature name into a string and then doing
regex-based decomposition.

The signature name is a particular (lossy) encoding of the
class/array decomposition of `ErasedTypeRef`, and of the
owner/name chain of `ClassSymbol`s. We now directly look at that
original information instead of recovering part of that information
from the signature name encoding.